Subject: Admin user should not be able to edit privileges or change role of a user
Description: We have implemented user authorization through LDAP. If a user has
write permissions on LDAP, the role of the user is set to Admin, to User
otherwise. So, user should not be able to manage these on the Users page.  This
commit removes the checkbox and button related to changing the privileges or
changing the role of the users.