Subject: When user goes to ADD A DOMAIN page and clicks Cancel at some point, page throws 400 error (PDIO-793)
Description: The cancel button did not work previously with javascript for on
click of the button and a valid url to dashboard.  This is because the button
is part of the entire input FORM element in the HTML with a POST action.
Before: The button's type was SUBMIT and redirect to a different url did not
work.  After: The button's type is changed to BUTTON and redirect to dashboard
url works.