Subject: Get New configuration for RBAC authorization for LDAP (PDIO-812)
Description:
This change is related to getting additional properties from the config file.
BaseDN and CommonName vary for this new LDAP query. User's RBAC permissions are
queried using these values.

Config file has the member value definition as either 'write_external' or
'write_internal'.  This value is set by puppet based on the hostname of the
box.

The code changes done as part of this commit are to get these new values from
the config file.