Subject: Implement LDAP authorization with RBAC integration (PDIO-812)
Description:
When a user requests for permissions on RBAC for DNS application, LDAP tree is
updated to have the permissions recorded for the user.

This code change queries the LDAP tree with the customized baseDN and
searchFilter with the common name.  Then, parses through the result and gets
the array of values under member in the tree.  Then, gets the member value set
in config file and verifies if any of the values in the array match this value.
If there is a match, give write permissions to the user. Else, read permission.
This is done by setting the role of the user to 'Admin' for write access and
'User' for reach access.