Subject: Performance related changes in update() of domains.
Description: This change is not related to any particular ticket.

Parts of the method update() have been commented to avoid the full sync of all
domains when the page loads.

Also, the method call in dashboard has been removed since we are not displaying
list of all domains anymore.

So, this change is not mandatory.

TODO : Remove this change and restore the original method definition after
thourough testing of the application.

The method does a full sync - deletes the domains in local db which have been
removed in the master, add new domains which are missing in local DB and exist
in master, update domain's details or records if anything changed in the
master.