Subject: In dashboard function, user get_domain doesn't work as domain user mapping isn't working currently.
Description: Domain User table doesn't get populated and domain user mapping
doesn't work.

Currently, this is not a requirement, so we will work on this later when we
need access level for each domain for a user.

Commenting the part where get_domain is called for a user. Instead, query all
from DB.

This is not really necessary because search.html rendering doesn't display all
the domains. But, I'm keeping the arguments of the html file the same
dashboard.html . i.e., pass domains array, domain_count, users, etc. This can
be cleaned up later.