Subject: Code fix for a bug in single domain sync
Description: Code changes in this commit are to put a fix for the following scenario:
Domain exists in local DB but not in API. Hence, API call for the domain throws error and the error gets caught in the exception.
Thus, not taking care of deletion in db. Domain always remains in DB even though its not there in API. Code change takes care of this deletion
inside the exception block.