Subject: Hide or View 'EDIT' and 'DELETE' buttons on domain page (PDIO-838)
Description: Code changes in this commit take care of hiding or viewing the edit and delete buttons for records of a domain.

Hide - if User is not Administrator or if domain is a Slave domain, then hide the buttons.
View - If user is an Administrator and the domain is Master/Native, then view the buttons.

Some of the code changes also take care of column number in the setup of user table. If the buttons are hidden then, its 5 columns else its 7.