Subject: In read view of zone clicking record changes the UI to show as editable even though it is not (PDIO-854)
Description:
For a slave zone or read permissions user, you can't edit the zone entries.
However, if you click on them the UI changes to make it look as though you can edit them,
even though you can't.
This change of appearance should probably not happen.

Changes made:

1. Fixed some indentation issues in domain.html

2. If domain type is slave or user has read only permissions, then the record rows
do not have clicking enabled.

There is a on click event defined in the jquery for the table rows which gets the edit mode.
So, disabling clicking if it's read view solves the issue.