Subject: For Search Domain page, add basic validations of the domain name (PDIO-856)
Description:

If domain name is empty, or contains spaces or special characters,
do not let the page talk to API and throw a 400 error.
Instead, validate on the UI and show a client side message.
The validation is similar to what's done on Create a new domain page.

Javascript function validateDomainName defines handling of different errors for domain name.

The function is called on click of Search button.