Subject: Power DNS GUI feedback: Identify the zone type (PDIO-957)
Description:
Currently there is no info displayed as to the zone type.
This makes it confusing as to why you don't have edit options on
certain zones (booking.com for example).
Displaying some form of warning/info to show the page is intended
to be view only because of xxx would be less confusing and result in
less questions from users.
Also displaying the relevant master ips etc for slaves would aid in
troubleshooting for anyone looking at the tool.