Subject: Staged changes are not clear (PDIO-958)
Description: Following changes are part of this commit:

1. Detect Unsaved Changes feature added. Any time user makes any change on the records,
if there is a diff between what's in the API and what's on the UI, a warning is displayed
asking the user to click Apply Changes to make the changes permanent.

2. If there are no changes and user clicked Apply Changes, a popup alert is displayed saying
there are no changes to be applied.

3. Once changes are applied through the API successfully, the page is refreshed so that
latest data is obtained from the API and the page is not stagnant with the data. Also, this helps
in clearning the unsaved changes warning message.