Subject: Diff Feature for Domain Edits
Description: Following changes are part of this commit:

1. domain.html - Added Show Diff Button and its bootstrap modal for diff details

2. domain.html - Changed the id for edit and delete buttons.

3. domain.html - Added global variables for edit diff, add diff, delete diff lists,
html before a row is edited and saved, last deleted row html dataTable index.

4. domain.html - Added unbind to button_delete_confirm modal to fix multiple firing
 of same event due to consecutive delete operations.

5. domain.html - In handle delete button function, if the deleted row doesn't exist in the
add diff list, add it to delete list.
Else remove it from add diff list, as its a change that cancels a previous change.

6. domain.html - In handle edit button function, keep track of the row html before it gets
edited. Store this in nBefore.

7. domain.html - In handle save, if its a new record, check if it's already present in delete
diff list and if so, remove it from delete diff as this is a change that cancels previous change.
If not, add it to add diff list.  NOTE: this check doesnt work because if an existing row is
deleted and then re-added, it loses its html index and the newly added row comes with a
totally new row index. So, the check always fails. Known BUG, have to fix it.

8. domain.html -  In handle show diff function, complete html for the modal is prepared.
This pulls data from edit, add, delete diff lists. This also takes care of
applying font colors of red and green for delete and add diff lists respectively.

9. domain.html - In detectUnsavedChanges function, previously entire table was compared.
In this commit, only the currently edited row is being compared to the server side value of
the record details. It's index is updated if any existing rows are deleted. This is done by
keeping track of lastDeletedIndex and computing the original server-side index of the
current record.

Also, this method detects what exactly changed and makes the font bold for that particular column.
And, if an existing record has changed, it updates the edit diff list.

NOTE: A loop is run for the server side Django variable records because there is no way to extract
one element using javascript index variable. As this is client-side variable. The only way is to check
each record from server-side and see if the index is same as the jaavscript index.
One of the cons of Django, I guess :(

10. domain.html - In htmlCleanUp function - the html that goes to show diff modal is updated.
Removes new lines, multiple spaces, empty columns, edit and delete buttons, etc.

11. domain.html - In setUnsavedChangesWarning function, a sanity check is done to keep what is seen
in show diff and what is seen as warning on main page , in sync. If all three diff lists are empty,
remove the warning.

12. custom.js - In saveRow function, id is updated for edit and delete buttons since its changed in domain.html.